artifacts/requirements.txt
PyYAML==6.0.2
httpx==0.27.2

